some base work for planes
